+++ instruments/swaps/interest_rate_swap.h
@@ -3,6 +3,7 @@
 
 #include "contract.h"
 #include "instruments/swaps/fixed_cashflow_instrument.h"
+#include "instruments/swaps/floating_cashflow_instrument.h"
 #include "rates/short_rate_tree_curve.h"
 
 namespace smileexplorer {
